datamesh-ai-connectors/aws-athena/requirements.txt
# AWS Athena Connector - Dependencies
# ====================================

# AWS SDK
boto3>=1.34.0
aiobotocore>=2.8.0

# Async support
aiohttp>=3.9.0

# Data handling
pandas>=2.1.0
pyarrow>=14.0.0

# Core datamesh dependencies
pydantic>=2.5.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
